+++ R/estimate_slopes.R
@@ -6,13 +6,14 @@
 #'  \item{\link[=estimate_slopes.stanreg]{Bayesian models (stanreg and brms)}}
 #'  }
 #'
-#' @inheritParams estimate_contrasts
-#' @param trend A character indicating the name of the numeric variable for which to compute the slopes.
-#' @param levels A character vectors indicating the variables over which the slope will be computed. If NULL (default), it will select all the remaining predictors.
+#' @inheritParams estimate_contrasts.lm
+#' @param trend A character vector indicating the name of the numeric variable for which to compute the slopes.
+#' @param levels A character vector indicating the variables over which the slope will be computed. If NULL (default), it will select all the remaining predictors.
+#' @param component A character vector indicating the model component for which estimation is requested. Only applies to models from \pkg{glmmTMB}. Use \code{"conditional"} for the count-model or \code{"zero_inflate"} or \code{"zi"} for the zero-inflation model.
 #'
-#' @return A dataframe of slopes.
+#' @return A data frame of slopes.
 #' @export
-estimate_slopes <- function(model, trend = NULL, levels = NULL, transform = "response", standardize = TRUE, standardize_robust = FALSE, ...) {
+estimate_slopes <- function(model, trend = NULL, levels = NULL, transform = "response", standardize = TRUE, standardize_robust = FALSE, ci = 0.95, ...) {
   UseMethod("estimate_slopes")
 }
 
@@ -44,7 +45,7 @@ estimate_slopes <- function(model, trend = NULL, levels = NULL, transform = "res
 #' }
 #' @importFrom stats mad median sd setNames
 #' @export
-estimate_slopes.stanreg <- function(model, trend = NULL, levels = NULL, transform = "response", standardize = TRUE, standardize_robust = FALSE, centrality = "median", ci = 0.89, ci_method = "hdi", test = c("pd", "rope"), rope_range = "default", rope_ci = 1, ...) {
+estimate_slopes.stanreg <- function(model, trend = NULL, levels = NULL, transform = "response", standardize = TRUE, standardize_robust = FALSE, ci = 0.95, centrality = "median", ci_method = "hdi", test = c("pd", "rope"), rope_range = "default", rope_ci = 1, ...) {
   .estimate_slopes(model, trend = trend, levels = levels, transform = transform, standardize = standardize, standardize_robust = standardize_robust, centrality = centrality, ci = ci, ci_method = ci_method, test = test, rope_range = rope_range, rope_ci = rope_ci)
 }
 
@@ -56,7 +57,7 @@ estimate_slopes.brmsfit <- estimate_slopes.stanreg
 #' Estimate the slopes of a numeric predictor (over different factor levels)
 #'
 #' @inheritParams estimate_slopes
-#' @inheritParams estimate_contrasts.stanreg
+#' @inheritParams estimate_contrasts.lm
 #'
 #' @examples
 #' library(modelbased)
@@ -65,12 +66,21 @@ estimate_slopes.brmsfit <- estimate_slopes.stanreg
 #' estimate_slopes(model)
 #' @export
 estimate_slopes.lm <- function(model, trend = NULL, levels = NULL, transform = "response", standardize = TRUE, standardize_robust = FALSE, ci = 0.95, ...) {
-  .estimate_slopes(model, trend = trend, levels = levels, transform = transform, standardize = standardize, standardize_robust = standardize_robust)
+  .estimate_slopes(model, trend = trend, levels = levels, transform = transform, standardize = standardize, standardize_robust = standardize_robust, ci=ci, ...)
 }
 
 
 #' @export
 estimate_slopes.merMod <- estimate_slopes.lm
+
+
+#' @rdname estimate_slopes
+#' @export
+estimate_slopes.glmmTMB <- function(model, trend = NULL, levels = NULL, transform = "response", standardize = TRUE, standardize_robust = FALSE, ci = 0.95, component = c("conditional", "zero_inflated", "zi"), ...) {
+  component <- match.arg(component)
+  if (component == "zi") component <- "zero_inflated"
+  .estimate_slopes(model, trend = trend, levels = levels, transform = transform, standardize = standardize, standardize_robust = standardize_robust, ci = ci, component = component, ...)
+}
 
 
 
@@ -82,12 +92,15 @@ estimate_slopes.merMod <- estimate_slopes.lm
 #' @importFrom stats confint
 #' @importFrom emmeans emtrends
 #' @keywords internal
-.estimate_slopes <- function(model, trend = NULL, levels = NULL, transform = "response", standardize = TRUE, standardize_robust = FALSE, centrality = "median", ci = 0.95, ci_method = "hdi", test = c("pd", "rope"), rope_range = "default", rope_ci = 1, ...) {
-  predictors <- insight::find_predictors(model)$conditional
+.estimate_slopes <- function(model, trend = NULL, levels = NULL, transform = "response", standardize = TRUE, standardize_robust = FALSE, ci = 0.95, centrality = "median", ci_method = "hdi", test = c("pd", "rope"), rope_range = "default", rope_ci = 1, component = "conditional", ...) {
+  predictors <- insight::find_predictors(model)[[component]]
   data <- insight::get_data(model)
 
   if (is.null(trend)) {
     trend <- predictors[sapply(data[predictors], is.numeric)][1]
+    if (!length(trend) || is.na(trend)) {
+      stop("Model contains no numeric predictor. Cannot estimate trend.")
+    }
     message('No numeric variable was specified for slope estimation. Selecting `trend = "', trend, '"`.')
   }
   if (length(trend) > 1) {
@@ -105,11 +118,7 @@ estimate_slopes.merMod <- estimate_slopes.lm
 
 
   # Basis
-  # Sometimes (when exactly?) fails when transform argument is passed
-  trends <- tryCatch(emmeans::emtrends(model, levels, var = trend, transform = transform, ...),
-    error = function(e) emmeans::emtrends(model, levels, var = trend, ...)
-  )
-
+  trends <- .emtrends_helper(model, levels, trend, transform, component, ...)
 
 
   if (insight::model_info(model)$is_bayesian) {
@@ -192,4 +201,35 @@ estimate_slopes.merMod <- estimate_slopes.lm
   }
   names(std) <- paste0("Std_", names(std))
   as.data.frame(std)
+}
+
+
+
+
+.emtrends_helper <- function(model, levels, trend, transform, component, ...) {
+  if (component != "conditional") {
+    if (transform == "response") {
+      emmeans::emtrends(model, levels, var = trend, transform = "response", component = "zi", ...)
+    } else if (transform == "mu") {
+      emmeans::emtrends(model, levels, var = trend, transform = "mu", component = "zi", ...)
+    } else if (transform == "unlink") {
+      emmeans::emtrends(model, levels, var = trend, transform = "unlink", component = "zi", ...)
+    } else if (transform == "log") {
+      emmeans::emtrends(model, levels, var = trend, transform = "log", component = "zi", ...)
+    } else if (transform == "none") {
+      emmeans::emtrends(model, levels, var = trend, transform = "none", component = "zi", ...)
+    }
+  } else {
+    if (transform == "response") {
+      emmeans::emtrends(model, levels, var = trend, transform = "response", ...)
+    } else if (transform == "mu") {
+      emmeans::emtrends(model, levels, var = trend, transform = "mu", ...)
+    } else if (transform == "unlink") {
+      emmeans::emtrends(model, levels, var = trend, transform = "unlink", ...)
+    } else if (transform == "log") {
+      emmeans::emtrends(model, levels, var = trend, transform = "log", ...)
+    } else if (transform == "none") {
+      emmeans::emtrends(model, levels, var = trend, transform = "none", ...)
+    }
+  }
 }